Ensure that users cannot implement `PinnedDrop` without proper attr
Fixes #179

Previously, it was possible to provide a `#[pinned_drop]` implementation
without annotating the pin-projected type with
`#[pin_project(PinnedDrop)]`. This would result in the `PinnedDrop` impl
never being used, without any warnings or errors being emitted to the
user.

This commit makes us generate a dummy `PinnedDrop` impl if the user does
not pass `PinnedDrop` as an argument to `#[pin_project]`. This dummy
impl will conflict with any user-provided impl, ensuring that the user
cannot accidentally a non-functional `PinnedDrop` impl.