Add `OriginalType<A, B>: 'pin` bound to projected type
Fixes #185

Usually, Rust will automatically add lifetime bounds of the form
`T: 'lifetime` when required (e.g. when a struct contains `&'a T`).

However, the compiler will not normalize associated types before
performing this analysis, which can require users to manually add
`T: 'lifetime` bounds themselves. When this happens with a generated
projection struct, the user has no ability to add such a bound, since
they cannot name the 'pin lifetime.

This PR adds a bound of the form `OriginalType<A, B>: 'pin` to the
`where` clause of the generated projection types. This ensures that
any type parameters which need to outlive 'pin will do so, without
unnecessarily constraining other type parameters.

The generated bound does not need to be added to the `where` clause of
any impls for the generated projection types. For more details, see
https://github.com/rust-lang/rfcs/blob/d89de6a5857597cca9f6e1c7ebb70476f31a9159/text/2093-infer-outlives.md#background-outlives-requirements-today