Allow clippy::redundant_pub_crate
When using pin-project within a module of a binary crate the
clippy::redundant_pub_crate (part of clippy::nursery) is triggered
due to the pub(crate) which is not needed in this case.

This change allows the clippy::redundant_pub_crate lint on all
generated code to silence the lint messages.